% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ada_boost.R
\name{AdaBoost}
\alias{AdaBoost}
\title{Demo of Adaboost using Sonar dataaset.}
\usage{
AdaBoost(train, test, index.x, index.y, n.iter)
}
\arguments{
\item{train}{training dataset.}

\item{test}{test dataset.}

\item{index.x}{index of X.}

\item{index.y}{index of Y.}

\item{n.iter}{the number of iterations.}
}
\value{
a list of objects:
   train.error errors from training dataset.
   test.error errors from test dataset.
}
\description{
The base learner is based on Recursive Partitioning and Regression Trees.
}
\seealso{
\url{http://web.stanford.edu/class/stats202/}
}
