% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_admm.R
\name{ADMM}
\alias{ADMM}
\title{ADMM alogrithm for solving convex optimization with l1-norm.}
\usage{
ADMM(rho, solATA, ATb, p, maxiter)
}
\arguments{
\item{maxiter.}{}
}
\value{
rrs residuals.
}
\description{
ADMM alogrithm for solving convex optimization with l1-norm.
}
\seealso{
\url{https://www.cs.cmu.edu/~ggordon/10725-F12/scribes/10725_Lecture19.pdf}
}
