% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvs_gibbs.R, R/bvs_spike_slab.R
\name{BvsGibbs}
\alias{BvsGibbs}
\alias{BvsGibbs}
\title{Gibbs sampler for Bayesian variables selection requires with continuous gaussian.}
\usage{
BvsGibbs(covariate, Y, nMc, v1, v0, w, gamma0, sigma0, a, b)

BvsGibbs(covariate, Y, nMc, v1, v0, w, gamma0, sigma0, a, b)
}
\arguments{
\item{covariate}{n by p covariate matrix (does not include the column of ones).}

\item{Y}{continuous response, n by 1.}

\item{nMc}{number of MCMC iterations.}

\item{v1}{p by 1 vector, the large s.d. of beta when gamma_j=1.}

\item{v0}{p by 1 vector, the small s.d. of beta when gamma_j=0.}

\item{w}{the probablity that gamma_j=1 for each variable.}

\item{gamma0}{initial value of gamma, contains zeros and ones.}

\item{sigma0}{initial value of sigma2.}

\item{a}{prior parameters for sigma^2. (Inverse Gamma parameters).}

\item{b}{prior parameters for sigma^2. (Inverse Gamma parameters).}

\item{beta0}{initial value of beta.}

\item{covariate}{n by p covariate matrix (does not include the column of ones).}

\item{Y}{continuous response, n by 1.}

\item{nMc}{number of MCMC iterations.}

\item{v1}{p by 1 vector, the large s.d. of beta when gamma_j=1.}

\item{v0}{p by 1 vector, the small s.d. of beta when gamma_j=0.}

\item{w}{the probablity that gamma_j=1 for each variable.}

\item{gamma0}{initial value of gamma, contains zeros and ones.}

\item{beta0}{initial value of beta.}

\item{sigma0}{initial value of sigma2.}

\item{a}{prior parameters for sigma^2. (Inverse Gamma parameters).}

\item{b}{prior parameters for sigma^2. (Inverse Gamma parameters).}
}
\value{
a list of objects:
   sample.beta All posterior samples of beta.
   sample.gamma all posterior samples of gamma.
   sample.sigma2 All posterior samples of sigma2.
   logodds The log (conditional) posterior odds used when updating gamma.

a list of objects:
   sample.beta All posterior samples of beta.
   sample.gamma all posterior samples of gamma.
   sample.sigma2 All posterior samples of sigma2.
   logodds The log (conditional) posterior odds used when updating gamma.
}
\description{
Gibbs sampler for Bayesian variables selection requires with continuous gaussian.

Gibbs sampler for Bayesian variables selection with spike-and-slab prior.
}
