% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvs_mh.R
\name{BvsMH}
\alias{BvsMH}
\title{Metropolis–Hastings algorithm for Bayesian variables selection with stochastic search.
For sampling scheme, we use switch and swap proposal (Brown, Vannucci, Fearn (1998), JRSS B).}
\usage{
BvsMH(covariate, Y, nMc, v1, v0, w, gamma0, a, b, phi)
}
\arguments{
\item{covariate}{n by p covariate matrix (does not include the column of ones).}

\item{Y}{continuous response, n by 1.}

\item{nMc}{number of MCMC iterations.}

\item{v1}{p by 1 vector, the large s.d. of beta when gamma_j=1.}

\item{v0}{p by 1 vector, the small s.d. of beta when gamma_j=0.}

\item{w}{the probablity that gamma_j=1 for each variable.}

\item{a}{prior parameters for sigma^2. (Inverse Gamma parameters).}

\item{b}{prior parameters for sigma^2. (Inverse Gamma parameters).}

\item{gamma0:initial}{value of tao, contains zeros and ones.}
}
\value{
sample.gamma all posterior samples of gamma
}
\description{
Metropolis–Hastings algorithm for Bayesian variables selection with stochastic search.
For sampling scheme, we use switch and swap proposal (Brown, Vannucci, Fearn (1998), JRSS B).
}
