% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coxph_backward_elimination.R
\name{bwe.cox}
\alias{bwe.cox}
\title{Backward feature elimination based on Wald type test for Cox model}
\usage{
bwe.cox(formula, data, alpha = 0.05, trace = 1, method = "breslow", ...)
}
\arguments{
\item{formula}{A model formula to start backward elimination}

\item{data}{A data frame}

\item{alpha}{Significant level. Default to 0.05}

\item{trace}{A integer for different levels of tracing. 0 = no trace, 1 = print dropped variable, 2 = print all considered models during the procedure}

\item{method}{Method to resolve tied event times pass to `coxph`. Default to 'breslow' instead of 'efron`}

\item{...}{Other parameters passed `coxph`}
}
\value{
A list with two entries: \cr
    $ fit: a `coxph` object of selected model
    $ wald.test: wald test p values for each feature and possible interaction.
}
\description{
Backward feature elimination based on Wald type test for Cox model
}
\details{
Notice that model hierachical structure is maintained, i.e., iff interactions involved in the model, its main effects will also be kept in the model.
}

