% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpr.R
\name{CalcSigma}
\alias{CalcSigma}
\title{Calculates the covariance matrix sigma using a the squared exponential function
with diagnal covraince matrix.}
\usage{
CalcSigma(X1, X2, l = 1)
}
\arguments{
\item{X1}{vectors of data.}

\item{X2}{vectors of data.}
}
\value{
Sigma a covariance matrix.
}
\description{
Calculates the covariance matrix sigma using a the squared exponential function
with diagnal covraince matrix.
}
