% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crp.R
\name{CRP}
\alias{CRP}
\title{Demo of chinese restaurant process.}
\usage{
CRP(num.customers, alpha)
}
\arguments{
\item{num.customers}{the number of customers (observaions).}

\item{alpha}{concentration parameter.}
}
\value{
Plot of clusetring results.
}
\description{
Demo of chinese restaurant process.
}
\seealso{
\url{https://github.com/tbroderick/bnp_tutorial/tree/2016mlss}
}
