% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_crp_gmm1d.R
\name{DpmmGibbs}
\alias{DpmmGibbs}
\title{Gibbs sampling for the Gaussian Dirichlet process mixture model.}
\usage{
DpmmGibbs(x, alpha, tau0, beta0, mu0, kappa0, nIter)
}
\arguments{
\item{x}{data (vector of length N).}

\item{alpha}{DP concentration parameter.}

\item{tau0}{normal-gamma prior shape.}

\item{beta0}{normal-gamma prior rate (inverse scale).}

\item{kappa0}{normal-gamma prior precision scaling parameter.}

\item{nIter}{number of Gibbs iterations.}
}
\value{
C N x nIter matrix of cluster assignments.
}
\description{
Gibbs sampling for the Gaussian Dirichlet process mixture model.
}
