% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_crp_gmm1d.R
\name{DpmmLoglik}
\alias{DpmmLoglik}
\title{Calculate the log likelihood for the Gaussian DP mixture model
Mean and variance parameters marginalized under normal-gamma prior
This corresponds to a generalized Student's t-distribution
NB: some constant terms are ignored}
\usage{
DpmmLoglik(xn, x, tau0, beta0, mu0, kappa0)
}
\description{
Calculate the log likelihood for the Gaussian DP mixture model
Mean and variance parameters marginalized under normal-gamma prior
This corresponds to a generalized Student's t-distribution
NB: some constant terms are ignored
}
