% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ELM_Survival.R
\name{elm}
\alias{elm}
\alias{elm.default}
\alias{elm.formula}
\alias{predict.elm}
\title{Extreme learning machine for survival, classification, regression (univariate / multivariate)}
\usage{
elm(x, ...)

\method{elm}{formula}(formula, data = environment(formula), ...)

\method{elm}{default}(x, y, nhid = max(200, 2 * ncol(x)), actfun = "sig",
  scaled = TRUE, family = switch(class(y), Surv = "cox", factor =
  ifelse(nlevels(y) > 2, "multinomial", "binomial"), integer = "poisson", matrix
  = "mgaussian", "gaussian"), lambda = ifelse(family == "mgaussian", 0, 1),
  ...)

\method{predict}{elm}(object, newdata, type = c("response", "probabilities",
  "link", "risk"))
}
\arguments{
\item{x}{Design matrix}

\item{...}{Further parameters passed to internal function}

\item{formula}{Model formula}

\item{data}{Training data frame}

\item{y}{Survival object, factor or a vector/matrix of continuous variable}

\item{nhid}{Number of (random) hidden neurons}

\item{actfun}{Activating function}

\item{scaled}{If to standardize input units}

\item{family}{Choices of 'cox', 'binomial', 'multinomial', 'poisson', 'gaussian', 'mgaussian'}

\item{lambda}{Shrinkage factor}

\item{object}{elm object}

\item{newdata}{A data frame (if elm called via formula) or a design matrix (if elm called via design matrix)}

\item{type}{Type of output}
}
\value{
elm object.
}
\description{
Extreme learning machine for survival, classification, regression (univariate / multivariate)
}
\examples{
library(survival);
data(pbc, package = 'randomForestSRC');
pbc <- na.omit(pbc);
i.tr <- sample(nrow(pbc), 100);

elm.f <- elm(Surv(days, status) ~., data = pbc[i.tr, ], nhid = 500);
elm.pred <- predict(object = elm.f, newdata = pbc[-i.tr, ], type = 'link');
survConcordance(Surv(days, status) ~ elm.pred, data = pbc[-i.tr, ])
}

