% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmean_faithful.R
\name{Euclid}
\alias{Euclid}
\title{Return Euclidean distance matrix for the data.}
\usage{
Euclid(points1, points2)
}
\arguments{
\item{points1}{data point.}

\item{points2}{data point.}
}
\value{
distanceMatrix D by D matrix.
}
\description{
Return Euclidean distance matrix for the data.
}
\seealso{
\url{http://stackoverflow.com/questions/31571236/my-own-k-means-algorithm-in-r}
}
