% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_mix_gauss_faithful.R
\name{GaussmixEM}
\alias{GaussmixEM}
\title{Demo of EM algorithm for univariate Gaussian Mixture Model.
using the Old Faithful dataset.}
\usage{
GaussmixEM(params, X, clusters = 2, tol = 1e-05, maxits = 100,
  showits = T)
}
\arguments{
\item{params:}{initial parameters (means, covariances, cluster probability).}

\item{X:}{data.}

\item{clusters:}{number of clusters desired.}

\item{tol:}{tolerance.}

\item{maxits:}{maximum iterations.}

\item{showits:}{whether to show iterations.}
}
\value{
a list of objects:
   probs: probability.
   mu: mean.
   var: variance.
   resp: responsibilities.
   cluster: cluster assignment.
}
\description{
Demo of EM algorithm for univariate Gaussian Mixture Model.
using the Old Faithful dataset.
}
