% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm_vbem.R
\name{GmmVbem}
\alias{GmmVbem}
\title{Variational Bayes EM algorithm for Gaussian Mixture Model.}
\usage{
GmmVbem(x, mix, PriorPar, options)
}
\arguments{
\item{x}{training data of size DxN.}

\item{mix}{gmm model initialize with netlab's GmmVBEM function.}

\item{PriorPar}{structure containing priors.}

\item{options}{options for maxIter, threshold etc. etc.}

\item{D}{the dimension.}

\item{N}{the number of Data points.}
}
\value{
a list of objects:
   L likelihood function.
   m centers of clusters.
}
\description{
Variational Bayes EM algorithm for Gaussian Mixture Model.
}
\seealso{
\url{http://www.cs.ubc.ca/~murphyk/Software/VBEMGMM/index.html}
}
