% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_sampling.R
\name{InverseCDF}
\alias{InverseCDF}
\title{Find inverse CDF}
\usage{
InverseCDF(F, l, u)
}
\arguments{
\item{F}{a cdf of f(x).}

\item{l}{lower bound of domain.}

\item{u}{upper bound of domain.}
}
\value{
F.inv inverse cdf.
}
\description{
Find inverse CDF
}
