% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/survival_kernel_knn.R
\name{kkm}
\alias{kkm}
\alias{kkm.default}
\alias{kkm.formula}
\title{Kernel Weighted Kaplan-Meier model for survival}
\usage{
kkm(x, ...)

\method{kkm}{formula}(formula, data, xtest = NULL, ytest = NULL, ...)

\method{kkm}{default}(x, y, xtest = NULL, ytest = NULL,
  return.train.prediction = FALSE || is.null(xtest), scaled = TRUE,
  k = min(30, nrow(x)), times = y[as.logical(y[, 2]), 1],
  type = c("kaplan-meier", "nelson-aalen"), kernel = "rbfdot",
  kpar = "automatic")
}
\arguments{
\item{x}{Design matrix (NO intercept)}

\item{...}{Further argument passed to internal functions}

\item{formula}{Model formula}

\item{data}{Data frame}

\item{xtest}{If the 'formula-data' format used, `xtest` is a data frame of the test set.
If the 'x-y' method called, `xtest` is the design matrix of the test set}

\item{ytest}{(optional) Survival outcome for testset}

\item{y}{Reponse vector of `Surv` object}

\item{scaled}{Logical value indicating if to standardize x, y}

\item{k}{Number of nearest neighbour used. Default to `nrow(x)` which seems the best}

\item{times}{Times to evaluate survival probabilities. Currently no used. All unique event times in training are used}

\item{type}{Type of estimator, either 'Kaplan-meier' or 'nelson-aalen'}

\item{kernel}{String or 'kernel' object (see kernlab::gausspr)}

\item{kpar}{A list of Kernel parameters or 'automatic' if a radial kernel specified}
}
\value{
A 'kkm' object
}
\description{
Kernel Weighted Kaplan-Meier model for survival
}
\examples{
library(survival);
data(pbc, package = 'randomForestSRC');
pbc <- na.omit(pbc);
i.tr <- sample(nrow(pbc), 100);
kkm.pred <- kkm(Surv(days, status) ~., data = pbc[i.tr, ], xtest = pbc[-i.tr, ], kernel = 'laplacedot', kpar = list(sigma = 0.1));
# concordance index if using 30th event time
survConcordance(Surv(days, status) ~ I(1 - kkm.pred$test.predicted.survival[, 30]), data = pbc[-i.tr, ])$concordance

plot(kkm.pred, subset = sample(length(i.tr), 10), lwd = 2)
}

