% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmean_faithful.R
\name{Kmeans}
\alias{Kmeans}
\title{Demo of k-Means algorithm using the Old Faithful dataset.}
\usage{
Kmeans(x, centers, dist.fun, nItter)
}
\arguments{
\item{x}{data.}

\item{centers}{centers of clusters.}

\item{dist.fun}{distance function l1/l2.}

\item{nItter}{number of iterations.}
}
\value{
a list of objects:
   clusters history of clusters
   centers history of centers.
}
\description{
Demo of k-Means algorithm using the Old Faithful dataset.
}
\seealso{
\url{http://www.cs.ubc.ca/~murphyk/Software/VBEMGMM/index.html}
}
