% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_gmm.R
\name{MetropolisHastingSampler}
\alias{MetropolisHastingSampler}
\title{Metropolis Hasting samlinging MCMC for 1-d Gaussian Mixture Model.}
\usage{
MetropolisHastingSampler(p, N)
}
\arguments{
\item{p}{target distribution.}

\item{N}{number of iterations need to run.}
}
\value{
xs samples.
}
\description{
Metropolis Hasting samlinging MCMC for 1-d Gaussian Mixture Model.
}
