% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_crp_gmm1d.R
\name{NormalizeLogprob}
\alias{NormalizeLogprob}
\title{Normalize a log probability vector without numerical underflow/overflow}
\usage{
NormalizeLogprob(log.prob)
}
\description{
Normalize a log probability vector without numerical underflow/overflow
}
