% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/survival_kernel_knn.R
\name{plot.kkm}
\alias{plot.kkm}
\title{Plot fitted survival curves}
\usage{
\method{plot}{kkm}(x, type = c("test", "train"), subset, xlab = "Time",
  ylab = "Survival", ...)
}
\arguments{
\item{x}{`kkm` object from `kkm`}

\item{type}{Either 'test' or 'train'}

\item{subset}{Subset of samples to plot, either a vector of logical or subscript/index}

\item{xlab}{Name of x label}

\item{ylab}{Name of y label}

\item{...}{Other parameters passed to `lattice::xyplot`}
}
\value{
A `trellis` object generated by `lattice::xyplot`
}
\description{
Plot fitted survival curves
}

