% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_urn.R
\name{PolyaUrn}
\alias{PolyaUrn}
\title{Demo of polya urn process.}
\usage{
PolyaUrn(G_0, N, alpha)
}
\arguments{
\item{G_0}{Base measure.}

\item{N}{the number of customers (observaions).}

\item{alpha}{concentration parameter.}
}
\value{
balls clusetring results.
}
\description{
Demo of polya urn process.
}
