% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rejection_sampling.R
\name{RejectionSampler}
\alias{RejectionSampler}
\title{Rejection sampler for univariate data.}
\usage{
RejectionSampler(theta.list, f, g, M)
}
\arguments{
\item{theta.list}{linspace.}

\item{f}{target distribution.}

\item{g}{proposed distribution.}

\item{M}{constant determinig the accptance rate.}
}
\value{
sample samples
}
\description{
Rejection sampler for univariate data.
}
