% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_linear_gauss.R
\name{SirLinearGauss}
\alias{SirLinearGauss}
\title{SMC method using SIR algorithm in linear Gaussian state-space model.}
\usage{
SirLinearGauss(y, T, N, S.v, S.w, phi)
}
\arguments{
\item{y}{obeservations.}

\item{T}{time.}

\item{N}{size.}

\item{S.v}{noise.}

\item{S.w}{noise.}

\item{phi}{scale parameter.}
}
\value{
a list of objects:
   mu.Xs mean sample.
   var.Xs variance sample.
}
\description{
SMC method using SIR algorithm in linear Gaussian state-space model.
}
