% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_stochastic_volatility.R
\name{SirStochasticVolatility}
\alias{SirStochasticVolatility}
\title{SMC method using SIR algorithm in linear Gaussian state-space model.}
\usage{
SirStochasticVolatility(y, T, N, gamma, sigma, phi)
}
\arguments{
\item{y}{obeservations.}

\item{T}{time.}

\item{N}{size.}

\item{sigma}{noise.}

\item{phi}{scale parameter.}
}
\value{
a list of objects:
   mu.Xs mean sample.
   var.Xs variance sample.
}
\description{
SMC method using SIR algorithm in linear Gaussian state-space model.
}
