\name{SML-internal}

\title{Internal SML functions}

\alias{check_isPosDouble}
\alias{check_isPosInteger}
\alias{check_isLogical}
\alias{check_isNumber}
\alias{generate_color}
\alias{generate_symbol}
\alias{generate_mixture_name}

\description{
  Internal functions not intended to be called directly by users.
}

\keyword{internal}
