\name{SML-package}
\alias{SML-package}
\alias{SML}
\docType{package}
\title{
Statistical Machine Learning
}
\description{
a collection of statistical machine learning methods.
}
\details{
\tabular{ll}{
Package: \tab SML\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-08-25\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Kevin Lin
Maintainers: Kevin Lin <klin1234@gmail.com>
}
\references{
1. Han Liu, Kathryn Roeder and Larry Wasserman. Stability Approach to Regularization Selection (StARS) for High Dimensional Graphical Models. \emph{Advances in Neural Information Processing Systems}(NIPS), 2010.\cr
}
\seealso{
\code{\link{lasso.stars}}
}