% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unigauss_vb.R
\name{UniGaussVb}
\alias{UniGaussVb}
\title{Univariate Variational Bayesian for 1-d Gaussian.}
\usage{
UniGaussVb(data, prior, post, Options, TrueNormalGammaPdf, vbPost)
}
\arguments{
\item{data}{data.}

\item{prior}{a list of priors.}

\item{post}{a list of posteriors.}

\item{Options}{a list of options.}

\item{TrueNormalGammaPdf}{true joint distribution.}

\item{vbPost}{estimated joint normal-gmma density.}
}
\value{
a list of objects:
   Lbound Lower bound of likelihood function.
   aN hyperparameter of lambda.
   bN hyperparameter of lambda.
   muN hyperparameter of mu.
   kappaN hyperparameter of mu.
}
\description{
Univariate Variational Bayesian for 1-d Gaussian.
}
